STUB

+++ sh/ghGenerate.sh
@@ -4,9 +4,9 @@ COMPRESS_CRIT="\.(ass|atom|bin|bm|bmp|conf|css|csv|htm|html|ico|js|json|kar|list
 #if [ ! -f "$(which zopfli)" ]; then
 	#sudo apt install -y zopfli
 #fi
-if [ ! -f "$(which tree)" ]; then
-	sudo apt install -y tree
-fi
+#if [ ! -f "$(which tree)" ]; then
+	#sudo apt install -y tree
+#fi
 
 rm -v 'book/FontAwesome/fonts/fontawesome-webfont.eot' 'book/FontAwesome/fonts/fontawesome-webfont.ttf'
 
@@ -44,7 +44,21 @@ find . -type f | while IFS= read -r file; do
 	if [ -f "$file" ]; then
 		# Is a file
 		if [ "$(echo "$file" | grep -E "$COMPRESS_CRIT")" != "" ]; then
-			rm "$file"
+			if [[ "$file" == *"/index.htm" ]]; then
+				#echo "File \"${file}\" is preserved."
+				printf "" > "$file"
+			elif [[ "$file" == *"/index.html" ]]; then
+				#echo "File \"${file}\" is preserved."
+				printf "" > "$file"
+			elif [[ "$file" == *"/default.htm" ]]; then
+				#echo "File \"${file}\" is preserved."
+				printf "" > "$file"
+			elif [[ "$file" == *"/default.html" ]]; then
+				#echo "File \"${file}\" is preserved."
+				printf "" > "$file"
+			else
+				rm "$file"
+			fi
 		else
 			echo "File \"${file}\" is preserved."
 		fi
